completed with titlebar of appication ui
