--- styles/Browser.module.css
@@ -1,170 +1,0 @@
-@media only screen and (max-width: 760px) {
-	.nav {
-		display: none;
-	}
-}
-
-.overlay {
-	background: radial-gradient(rgba(0, 0, 0, 0.5), rgba(250, 250, 250, 0.5));
-	background-blend-mode: luminosity;
-	z-index: 2;
-	border-radius: 10px;
-}
-
-@media only screen and (min-width: 760px) {
-	.browser {
-		height: 90vh;
-		width: 90vw;
-		border: 3px solid black;
-		display: flex;
-		flex-direction: column;
-		border-radius: 10px;
-		overflow: hidden;
-		box-shadow: 0px 0px 0.1em #fff;
-		animation: grow 0.5s ease-in-out;
-	}
-	.nav {
-		height: 2.5em;
-		width: 100%;
-		background-color: rgba(44, 64, 112, 0.7);
-		margin-top: 0;
-		padding: 0px 10px;
-		display: flex;
-		overflow: hidden;
-	}
-	.nav ul {
-		display: flex;
-		margin: 0;
-		padding: 0;
-		height: 100%;
-		flex: 5;
-	}
-	.nav li {
-		height: 100%;
-		max-width: 10rem;
-		background-color: rgba(44, 64, 112, 0.9);
-		font-size: 1.1em;
-		font-weight: 300;
-		padding-top: 0.5rem;
-		border-bottom-left-radius: 0px;
-		border-bottom-right-radius: 0px;
-		display: flex;
-		flex: 1;
-		justify-content: space-between;
-		text-align: center;
-		text-transform: capitalize;
-		color: #fff;
-		box-shadow: 1px -1px 1px rgba(250, 250, 250, 0.9);
-	}
-	.nav .isActive {
-		border: 0.1px solid rgb(160, 160, 255);
-		background-color: rgba(13, 14, 34, 0.9);
-		border-top: 0;
-		border-bottom: 0;
-	}
-	.nav li:nth-child(1) {
-		border-top-left-radius: 10px;
-		color: #000;
-	}
-	/* .nav li:last-child {
-		border-top-right-radius: 10px;
-	} */
-	.nav li:last-child::after {
-		content: "";
-		display: none;
-		border-top-right-radius: 10px;
-	}
-	.nav li::after {
-		content: "";
-		height: 100%;
-		width: 1px;
-		background-color: #fff;
-		align-self: center;
-		justify-self: center;
-		margin-bottom: 5px;
-	}
-	.nav li a {
-		display: flex;
-	}
-	.nav .notActive:hover {
-		background-color: rgba(13, 14, 34, 0.5);
-	}
-	.navIcon {
-		height: 1.5rem;
-		margin: 0px 10px;
-	}
-
-	.content {
-		height: Calc(100% - 2.5em);
-	}
-	.cancel {
-		justify-self: end;
-		flex: 0.1;
-		margin: 0;
-		padding: 0.5em;
-		cursor: pointer;
-		position: relative;
-	}
-	.cancel::before,
-	.cancel::after {
-		content: "";
-		width: 3px;
-		height: 1.5rem;
-		background: red;
-		position: absolute;
-		right: 50%;
-		top: 25%;
-	}
-	.cancel::after {
-		transform: rotate(45deg);
-	}
-	.cancel::before {
-		transform: rotate(-45deg);
-	}
-	.cancel:hover,
-	.cancel:focus {
-		background: red;
-	}
-	.cancel:hover::before,
-	.cancel:hover::after {
-		opacity: 0.8;
-		background: #fff;
-		transition: opacity 250ms linear;
-	}
-
-	.background {
-		color: #fff;
-		font-weight: 600;
-		animation: shrink 0.3s linear;
-	}
-	.background p {
-		margin: 3px;
-	}
-	.backIcon {
-		height: auto;
-		width: auto;
-		margin: 5px;
-		filter: drop-shadow(0px 0px 5px rgba(37, 187, 168, 0.7));
-	}
-	.background:hover {
-		filter: drop-shadow(0px 0px 5px rgba(37, 187, 168, 0.7));
-	}
-}
-
-@keyframes grow {
-	from {
-		transform: scale(0);
-		overflow: hidden;
-	}
-	to {
-		transform: scale(1);
-	}
-}
-@keyframes shrink {
-	from {
-		transform: scale(2);
-	}
-	to {
-		transform: scale(1);
-	}
-}
